Executor/Driver Pattern, Simplified Builds, VFS
This patch introduces a simplified executor/driver design pattern,
simplified builds, and a VFS executor/driver for testing.
