+++ .gitignore
@@ -1,9 +1,12 @@
+*.a
+*.d
+*.out
+*.lock
 libstorage.paw
 .site/
 site/
 vendor/
 .project
-glide.lock
 .gomk/tmp
 api/server/executors/executors_generated.go
