--- .gomk/include/tools/deploy/bintray.mk
@@ -1,8 +1,0 @@
-ifneq (1,$(IS_GOMK_BINTRAY_LOADED))
-
-# note that the file is loaded
-IS_GOMK_BINTRAY_LOADED := 1
-
-
-
-endif
