--- api/server/executors/executors.go
@@ -1,1 +1,0 @@
-package executors
