--- api/server/executors/gomk.properties
@@ -1,75 +1,0 @@
-
-define LIBSTORAGE_EXECUTORS_XBUILD_DEF
-
-LSEXB_X_BP_$1_$2 := $$(subst -, ,$2)
-LSEXB_X_BP_OS_$1_$2 := $$(firstword $$(LSEXB_X_BP_$1_$2))
-LSEXB_X_BP_ARCH_$1_$2 := $$(lastword $$(LSEXB_X_BP_$1_$2))
-
-ifeq (Linux,$$(LSEXB_X_BP_OS_$1_$2))
-	LSEXB_X_GOOS_$1_$2 := linux
-else
-	ifeq (Darwin,$$(LSEXB_X_BP_OS_$1_$2))
-		LSEXB_X_GOOS_$1_$2 := darwin
-	else
-		ifeq (Windows,$$(LSEXB_X_BP_OS_$1_$2))
-			LSEXB_X_GOOS_$1_$2 := windows
-		endif
-	endif
-endif
-
-ifeq (i386,$$(LSEXB_X_BP_ARCH_$1_$2))
-	LSEXB_X_GOARCH_$1_$2 := 386
-else
-	ifeq (x86_64,$$(LSEXB_X_BP_ARCH_$1_$2))
-		LSEXB_X_GOARCH_$1_$2 := amd64
-	endif
-endif
-
-ifeq (1,$(RECURSIVE))
-	RECURSIVE_INDENT := INDENT_LEN=$$(INDENT_LEN)
-endif
-
-$1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2) := $$(GOMK_TMP_DIR)/$1_executor/build/bin/$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2)/libstorage-$1-executor
-$1-executor-build-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2): $$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))
-$$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2)): | $$(ENV)
-	$$(ENV) $$(RECURSIVE_INDENT) \
-		GOMK_TOOLS_ENABLE=1 \
-		GO_TAGS='$1 executor' \
-		LIBSTORAGE_EXECUTOR_DRIVER=$1 \
-		BUILD_EXECUTOR=1 \
-		GOOS=$$(LSEXB_X_GOOS_$1_$2) \
-		GOARCH=$$(LSEXB_X_GOARCH_$1_$2) \
-		$$(MAKE) build-executor
-
-$$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))-clean: | $$(RM)
-	$(RM) -f $$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))
-GO_PHONY += $$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))-clean
-GO_CLEAN += $$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))-clean
-
-LIBSTORAGE_EXECUTORS += $$($1-executor-$$(LSEXB_X_GOOS_$1_$2)_$$(LSEXB_X_GOARCH_$1_$2))
-
-endef
-
-define LIBSTORAGE_EXECUTORS_DEF
-$$(foreach bp,$$(BUILD_PLATFORMS),$$(eval $$(call LIBSTORAGE_EXECUTORS_XBUILD_DEF,$1,$$(bp))))
-endef
-
-ifneq (1,$(BUILD_EXECUTOR))
-$(foreach d,$(DRIVERS),$(eval $(call LIBSTORAGE_EXECUTORS_DEF,$(d))))
-endif
-
-LIBSTORAGE_EXECUTORS_EMBEDDED := ./api/server/executors/executors_generated.go
-build-embedded-executors: $(LIBSTORAGE_EXECUTORS_EMBEDDED)
-$(LIBSTORAGE_EXECUTORS_EMBEDDED): $(LIBSTORAGE_EXECUTORS) | $(GOBINDATA)
-	$(GOBINDATA) -tags '$(DRIVERS)' \
-                 -pkg executors \
-                 -o $@ \
-                 -prefix "$(GO_TMP_BUILD_DIR)/bin/" \
-                 $(dir $(LIBSTORAGE_EXECUTORS))
-$(LIBSTORAGE_EXECUTORS_EMBEDDED)-clean: | $(RM)
-	$(RM) -f $(LIBSTORAGE_EXECUTORS_EMBEDDED)
-
-GO_CLEAN += $(LIBSTORAGE_EXECUTORS_EMBEDDED)-clean
-GO_PHONY += $(LIBSTORAGE_EXECUTORS_EMBEDDED)-clean
-
-GO_PKG_BUILD_DEPS_./api/server/executors += $(LIBSTORAGE_EXECUTORS_EMBEDDED)
