--- api/server/gomk.properties
@@ -1,2 +1,0 @@
-GO_PKG_BUILD_DEPS_./api/server += ./api/server/executors/executors_generated.go
-GO_PKG_BUILD_DEPS_./api/server += $(GO_TMP_BUILD_PKG_DIR)/libstorage/api/server/executors.a
