--- api/server/router/driver/driver_ignore.go
@@ -1,1 +1,0 @@
-package driver
