+++ api/server/server.go
@@ -21,6 +21,9 @@ import (
 	// imported to load routers
 	_ "github.com/emccode/libstorage/api/server/router"
 
+	// imported to load drivers
+	_ "github.com/emccode/libstorage/drivers"
+
 	"github.com/emccode/libstorage/api/registry"
 	"github.com/emccode/libstorage/api/server/httputils"
 	"github.com/emccode/libstorage/api/types/context"
