--- api/utils/schema/gomk.properties
@@ -1,9 +1,0 @@
-LIBSTORAGE_JSON := ./libstorage.json
-LIBSTORAGE_JSON_EMBEDDED := ./api/utils/schema/schema_generated.go
-
-$(LIBSTORAGE_JSON_EMBEDDED):: $(LIBSTORAGE_JSON) | $(PRINTF) $(SED)
-	@$(PRINTF) "package schema\n\nconst (\n" >$@
-	@$(PRINTF) "\t// JSONSchema is the libStorage API JSON schema\n" >>$@
-	@$(PRINTF) "\tJSONSchema = \`" >>$@
-	@$(SED) -e 's/^//' $< >>$@
-	@$(PRINTF) "\`\n)\n" >>$@
