--- cli/executor/gomk.properties
@@ -1,8 +1,0 @@
-ifeq (,$(LIBSTORAGE_EXECUTOR_DRIVER))
-LIBSTORAGE_EXECUTOR_DRIVER := mock
-endif
-
-ifeq (1,$(BUILD_EXECUTOR))
-GO_BUILD_OUTPUT_FILE_./cli/executor := $(GO_TMP_BUILD_BIN_DIR)/libstorage-$(LIBSTORAGE_EXECUTOR_DRIVER)-executor
-build-executor: $(GO_BUILD_OUTPUT_FILE_./cli/executor)
-endif
