+++ cli/executor/libstorage-isilon-executor/executor.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/emccode/libstorage/cli/executor"
+)
+
+func main() {
+	executor.Run()
+}
