+++ cli/executor/libstorage-mock-executor/executor.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"github.com/emccode/libstorage/cli/executor"
+
+	// load the executor
+	_ "github.com/emccode/libstorage/drivers/storage/mock/executor"
+)
+
+func main() {
+	executor.Run()
+}
