+++ drivers/storage/mock/mock_driver.go
@@ -1,7 +1,3 @@
-// +build mock
-// +build driver
-// +build !executor
-
 package mock
 
 import (
@@ -11,11 +7,98 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/akutz/goof"
 
+	"github.com/emccode/libstorage/api/registry"
 	"github.com/emccode/libstorage/api/types"
 	"github.com/emccode/libstorage/api/types/context"
 	"github.com/emccode/libstorage/api/types/drivers"
 	"github.com/emccode/libstorage/api/utils"
+	"github.com/emccode/libstorage/drivers/storage/mock/executor"
 )
+
+type driver struct {
+	executor.Executor
+	nextDeviceInfo *types.NextDeviceInfo
+	volumes        []*types.Volume
+	snapshots      []*types.Snapshot
+	storageType    types.StorageType
+}
+
+func init() {
+	registry.RegisterStorageDriver(executor.Name1, newDriver1)
+	registry.RegisterStorageDriver(executor.Name2, newDriver2)
+	registry.RegisterStorageDriver(executor.Name3, newDriver3)
+}
+
+func newDriver1() drivers.StorageDriver {
+	return newDriver(executor.Name1)
+}
+
+func newDriver2() drivers.StorageDriver {
+	return newDriver(executor.Name2)
+}
+
+func newDriver3() drivers.StorageDriver {
+	return newDriver(executor.Name3)
+}
+
+func newDriver(name string) drivers.StorageDriver {
+
+	d := &driver{Executor: *executor.NewExecutor(name)}
+
+	d.nextDeviceInfo = &types.NextDeviceInfo{
+		Prefix:  "xvd",
+		Pattern: `\w`,
+		Ignore:  getDeviceIgnore(name),
+	}
+
+	d.volumes = []*types.Volume{
+		&types.Volume{
+			Name:             d.pwn("Volume 0"),
+			ID:               d.pwn("vol-000"),
+			AvailabilityZone: d.pwn("zone-000"),
+			Type:             "gold",
+			Size:             10240,
+		},
+		&types.Volume{
+			Name:             d.pwn("Volume 1"),
+			ID:               d.pwn("vol-001"),
+			AvailabilityZone: d.pwn("zone-001"),
+			Type:             "gold",
+			Size:             40960,
+		},
+		&types.Volume{
+			Name:             d.pwn("Volume 2"),
+			ID:               d.pwn("vol-002"),
+			AvailabilityZone: d.pwn("zone-002"),
+			Type:             "gold",
+			Size:             163840,
+		},
+	}
+
+	d.snapshots = []*types.Snapshot{
+		&types.Snapshot{
+			Name:     d.pwn("Snapshot 0"),
+			ID:       d.pwn("snap-000"),
+			VolumeID: d.pwn("vol-000"),
+		},
+		&types.Snapshot{
+			Name:     d.pwn("Snapshot 1"),
+			ID:       d.pwn("snap-001"),
+			VolumeID: d.pwn("vol-001"),
+		},
+		&types.Snapshot{
+			Name:     d.pwn("Snapshot 2"),
+			ID:       d.pwn("snap-002"),
+			VolumeID: d.pwn("vol-002"),
+		},
+	}
+
+	return d
+}
+
+func (d *driver) Type() types.StorageType {
+	return types.Block
+}
 
 func (d *driver) NextDeviceInfo() *types.NextDeviceInfo {
 	return d.nextDeviceInfo
@@ -24,7 +107,8 @@ func (d *driver) NextDeviceInfo() *types.NextDeviceInfo {
 func (d *driver) InstanceInspect(
 	ctx context.Context,
 	opts types.Store) (*types.Instance, error) {
-	return &types.Instance{InstanceID: d.instanceID}, nil
+	iid, _ := d.InstanceID(ctx, opts)
+	return &types.Instance{InstanceID: iid}, nil
 }
 
 func (d *driver) Volumes(
@@ -287,4 +371,39 @@ func (d *driver) SnapshotRemove(
 	d.snapshots = append(d.snapshots[:xToRemove], d.snapshots[xToRemove+1:]...)
 
 	return nil
+}
+
+func (d *driver) pwn(v string) string {
+	return fmt.Sprintf("%s-%s", d.Name(), v)
+}
+
+func getDeviceIgnore(driver string) bool {
+	if driver == executor.Name2 {
+		return true
+	}
+	return false
+}
+
+func getDeviceName(driver string) string {
+	var deviceName string
+	switch driver {
+	case executor.Name1:
+		deviceName = "/dev/xvdb"
+	case executor.Name2:
+		deviceName = "/dev/xvda"
+	case executor.Name3:
+		deviceName = "/dev/xvdc"
+	}
+	return deviceName
+}
+
+func getNextDeviceName(driver string) string {
+	var deviceName string
+	switch driver {
+	case executor.Name1:
+		deviceName = "/dev/xvdc"
+	case executor.Name3:
+		deviceName = "/dev/xvdb"
+	}
+	return deviceName
 }
