--- drivers/storage/mock/mock_nodriver.go
@@ -1,109 +1,0 @@
-// +build mock
-// +build !driver
-
-package mock
-
-import (
-	"github.com/emccode/libstorage/api/types"
-	"github.com/emccode/libstorage/api/types/context"
-	"github.com/emccode/libstorage/api/types/drivers"
-)
-
-func (d *driver) NextDeviceInfo() *types.NextDeviceInfo {
-	return nil
-}
-
-func (d *driver) InstanceInspect(
-	ctx context.Context,
-	opts types.Store) (*types.Instance, error) {
-	return nil, nil
-}
-
-func (d *driver) Volumes(
-	ctx context.Context,
-	opts *drivers.VolumesOpts) ([]*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeInspect(
-	ctx context.Context,
-	volumeID string,
-	opts *drivers.VolumeInspectOpts) (*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeCreate(
-	ctx context.Context,
-	name string,
-	opts *drivers.VolumeCreateOpts) (*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeCreateFromSnapshot(
-	ctx context.Context,
-	snapshotID, volumeName string,
-	opts types.Store) (*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeCopy(
-	ctx context.Context,
-	volumeID, volumeName string,
-	opts types.Store) (*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeSnapshot(
-	ctx context.Context,
-	volumeID, snapshotName string,
-	opts types.Store) (*types.Snapshot, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeRemove(
-	ctx context.Context,
-	volumeID string,
-	opts types.Store) error {
-	return nil
-}
-
-func (d *driver) VolumeAttach(
-	ctx context.Context,
-	volumeID string,
-	opts *drivers.VolumeAttachByIDOpts) (*types.Volume, error) {
-	return nil, nil
-}
-
-func (d *driver) VolumeDetach(
-	ctx context.Context,
-	volumeID string,
-	opts types.Store) error {
-	return nil
-}
-
-func (d *driver) Snapshots(
-	ctx context.Context,
-	opts types.Store) ([]*types.Snapshot, error) {
-	return nil, nil
-}
-
-func (d *driver) SnapshotInspect(
-	ctx context.Context,
-	snapshotID string,
-	opts types.Store) (*types.Snapshot, error) {
-	return nil, nil
-}
-
-func (d *driver) SnapshotCopy(
-	ctx context.Context,
-	snapshotID, snapshotName, destinationID string,
-	opts types.Store) (*types.Snapshot, error) {
-	return nil, nil
-}
-
-func (d *driver) SnapshotRemove(
-	ctx context.Context,
-	snapshotID string,
-	opts types.Store) error {
-	return nil
-}
