--- drivers/storage/mock/mock_noop.go
@@ -1,3 +1,0 @@
-// +build !mock
-
-package mock
