+++ drivers/storage/storage.go
@@ -3,4 +3,5 @@ package storage
 import (
 	// import to load
 	_ "github.com/emccode/libstorage/drivers/storage/mock"
+	_ "github.com/emccode/libstorage/drivers/storage/vfs"
 )
