+++ drivers/storage/vfs/executor/vfs_executor_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package executor
+
+func getHostName() (string, error) {
+	return "windows", nil
+}
