--- libstorage_nomock_test.go
@@ -1,13 +1,0 @@
-// +build !mock
-
-package libstorage
-
-import (
-	"testing"
-
-	"github.com/akutz/gofig"
-)
-
-func getConfig(host string, tls bool, t *testing.T) gofig.Config {
-	return gofig.New()
-}
